+++ src/components/game/MainHeader.tsx
@@ -8,13 +8,17 @@ import { motion, AnimatePresence } from 'framer-motion';
 import { Home, Map, MessageSquare, Menu, X } from 'lucide-react';
 import { useExerciseTracking } from '@/context/ExerciseTrackingContext';
 
+interface MainHeaderProps {
+  headerActions?: React.ReactNode; 
+}
+
 const navItems = [
   { href: '/home', icon: <Home className="w-5 h-5" />, label: 'Home' },
   { href: '/map', icon: <Map className="w-5 h-5" />, label: 'Mapa' },
   { href: '/chat', icon: <MessageSquare className="w-5 h-5" />, label: 'Chat AI' },
 ];
 
-export default function MainHeader() {
+export default function MainHeader({ headerActions }: MainHeaderProps) {
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
   const router = useRouter();
   const { trackInteraction } = useExerciseTracking();
@@ -33,14 +37,16 @@ export default function MainHeader() {
 
   return (
     <>
-      <header className="fixed top-0 left-0 right-0 z-50 bg-white/80 backdrop-blur-md border-b border-gray-200/80">
+      <header id="main-header" className="fixed top-0 left-0 right-0 z-50 bg-white backdrop-blur-md border-b border-gray-200/80">
         <nav className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
           <div className="flex items-center justify-between h-16">
             {/* Logo / Título */}
             <Link href="/" className="flex items-center gap-2 text-xl font-bold text-gray-800">
               <span className="w-8 h-8 bg-gradient-to-r from-blue-600 to-indigo-600 rounded-lg flex items-center justify-center text-white font-mono text-lg">AI</span>
               FluentBlocks
+
             </Link>
+            {headerActions && <div className="p-2 border-b border-gray-100">{headerActions}</div>}
 
             {/* Navegación de Escritorio */}
             <div className="hidden md:flex items-center space-x-2">
